Extract AnnotatorScanner config writer into a separate class. (#116)
This PR simply extracts the config writer inside the `annotator-scanner` module into a separate standalone class, and reuses it in `annotator-core` module.
