+++ annotator-scanner/src/main/java/edu/ucr/cs/riple/scanner/Config.java
@@ -24,19 +24,8 @@
 
 package edu.ucr.cs.riple.scanner;
 
-import com.google.common.base.Preconditions;
 import java.nio.file.Path;
 import javax.annotation.Nonnull;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerException;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
 
 /** Config of scanner. */
 public interface Config {
@@ -83,103 +72,4 @@ public interface Config {
    */
   @Nonnull
   Path getOutputDirectory();
-
-  /** Builder for setting a AnnotatorScanner configuration and output in XML format. */
-  class Builder {
-
-    /** Path to output directory. */
-    private Path outputDirectory;
-    /** Controls method info serialization. */
-    private boolean methodTrackerIsActive;
-    /** Controls field usage info serialization. */
-    private boolean fieldTrackerIsActive;
-    /** Controls method invocation serialization. */
-    private boolean callTrackerIsActive;
-    /** Controls class info serialization. */
-    private boolean classTrackerIsActive;
-
-    public Builder() {
-      this.methodTrackerIsActive = false;
-      this.fieldTrackerIsActive = false;
-      this.callTrackerIsActive = false;
-      this.classTrackerIsActive = false;
-    }
-
-    public Builder setOutput(Path output) {
-      this.outputDirectory = output;
-      return this;
-    }
-
-    public Builder setMethodTrackerActivation(boolean activation) {
-      this.methodTrackerIsActive = activation;
-      return this;
-    }
-
-    public Builder setFieldTrackerActivation(boolean activation) {
-      this.fieldTrackerIsActive = activation;
-      return this;
-    }
-
-    public Builder setCallTrackerActivation(boolean activation) {
-      this.callTrackerIsActive = activation;
-      return this;
-    }
-
-    public Builder setClassTrackerActivation(boolean activation) {
-      this.classTrackerIsActive = activation;
-      return this;
-    }
-
-    /**
-     * Outputs the configured object as XML format in the given path.
-     *
-     * @param path Output path.
-     */
-    public void writeAsXML(Path path) {
-      Preconditions.checkNotNull(this.outputDirectory, "Output directory must be initialized.");
-      DocumentBuilderFactory docFactory = DocumentBuilderFactory.newInstance();
-      try {
-        DocumentBuilder docBuilder = docFactory.newDocumentBuilder();
-        Document doc = docBuilder.newDocument();
-
-        // Root
-        Element rootElement = doc.createElement("scanner");
-        doc.appendChild(rootElement);
-
-        // Method
-        Element methodElement = doc.createElement("method");
-        methodElement.setAttribute("active", String.valueOf(methodTrackerIsActive));
-        rootElement.appendChild(methodElement);
-
-        // Field
-        Element fieldElement = doc.createElement("field");
-        fieldElement.setAttribute("active", String.valueOf(fieldTrackerIsActive));
-        rootElement.appendChild(fieldElement);
-
-        // Call
-        Element callElement = doc.createElement("call");
-        callElement.setAttribute("active", String.valueOf(callTrackerIsActive));
-        rootElement.appendChild(callElement);
-
-        // File
-        Element classElement = doc.createElement("class");
-        classElement.setAttribute("active", String.valueOf(classTrackerIsActive));
-        rootElement.appendChild(classElement);
-
-        // Output dir
-        Element outputDir = doc.createElement("path");
-        outputDir.setTextContent(this.outputDirectory.toString());
-        rootElement.appendChild(outputDir);
-
-        // Writings
-        TransformerFactory transformerFactory = TransformerFactory.newInstance();
-        Transformer transformer = transformerFactory.newTransformer();
-        DOMSource source = new DOMSource(doc);
-        StreamResult result = new StreamResult(path.toFile());
-        transformer.transform(source, result);
-      } catch (ParserConfigurationException | TransformerException e) {
-        throw new RuntimeException("Error happened in writing config.", e);
-      }
-    }
-  }
 }
