+++ annotator-scanner/src/test/java/edu/ucr/cs/riple/scanner/AnnotatorScannerBaseTest.java
@@ -56,18 +56,17 @@ public abstract class AnnotatorScannerBaseTest<T extends Display> {
   @Before
   public void setup() {
     root = Paths.get(temporaryFolder.getRoot().getAbsolutePath());
-    Path config = root.resolve("scanner.xml");
+    Path configPath = root.resolve("scanner.xml");
     try {
       Files.createDirectories(root);
-      ErrorProneCLIFlagsConfig.Builder builder =
-          new ErrorProneCLIFlagsConfig.Builder()
-              .setCallTrackerActivation(true)
-              .setClassTrackerActivation(true)
-              .setFieldTrackerActivation(true)
-              .setMethodTrackerActivation(true)
-              .setOutput(root);
-      Files.createFile(config);
-      builder.writeAsXML(config);
+      ScannerConfigWriter writer = new ScannerConfigWriter();
+      writer
+          .setCallTrackerActivation(true)
+          .setClassTrackerActivation(true)
+          .setFieldTrackerActivation(true)
+          .setMethodTrackerActivation(true)
+          .setOutput(root)
+          .writeAsXML(configPath);
     } catch (IOException ex) {
       throw new UncheckedIOException(ex);
     }
@@ -78,7 +77,7 @@ public abstract class AnnotatorScannerBaseTest<T extends Display> {
                     "-d",
                     temporaryFolder.getRoot().getAbsolutePath(),
                     "-Xep:AnnotatorScanner:ERROR",
-                    "-XepOpt:AnnotatorScanner:ConfigPath=" + config))
+                    "-XepOpt:AnnotatorScanner:ConfigPath=" + configPath))
             .setOutputFileNameAndHeader(fileName, header)
             .setFactory(factory);
   }
